pyproject.toml
[tool.poetry]
name = "jenkins-script"
version = "0.1.0"
description = ""
authors = ["imadenugraha <imadenugraha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
python-jenkins = "^1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
